less automatic with symetric matrices, synthe data generation enhanced, still under work
git-svn-id: file:///home/galbrun/TKTL/redescriptors/old_svn/redescriptions-svn/sandbox@57 8b04a84e-e702-47ad-8570-c905424f7f45

+++ scripts/utilsIO.py
@@ -89,7 +89,7 @@ def readMatlabNum(filename):
         a = row.split('\t')
         id_row = int(a[0])
         id_col = int(a[1])
-        val = int(a[2])
+        val = float(a[2])
         ## For Python 2.4, convert a to INT
         if tmpcolSupps == []:
             tmpcolSupps = [[(0, -1)] for i in range(id_col)]
